requirements.txt
pika==1.3.1
aiohttp==3.8.5
lxml==4.9.3
python-dotenv==1.0.0
